New Structure, start from the beginning in empty fill
