+++ Gamesold/Avalon/avalon-function.js
