+++ Gamesold/GameTemplate/gameTemplate-commands.js
