+++ Gamesold/GameTemplate/gameTemplate.js
